+++ app/src/main/java/woowacourse/paint/customview/paint/PaintMode.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.customview
+package woowacourse.paint.customview.paint
 
 enum class PaintMode {
     PEN, RECTANGLE, OVAL, ERASER
