+++ app/src/main/java/woowacourse/paint/customview/paint/tool/Eraser.kt
@@ -1,16 +1,17 @@
-package woowacourse.paint.customview
+package woowacourse.paint.customview.paint.tool
 
 import android.graphics.Paint
 import android.graphics.Path
+import android.graphics.PorterDuff
+import android.graphics.PorterDuffXfermode
+import woowacourse.paint.customview.paint.Painting
 
-class Pen(private val penColor: Int, private val thickness: Float) : PaintTool {
-
+class Eraser(private val thickness: Float) : PaintTool {
     override val painting: Painting = Painting(Path(), initPaint())
 
     private fun initPaint(): Paint {
         return Paint().apply {
-            color = penColor
-            isAntiAlias = true
+            xfermode = PorterDuffXfermode(PorterDuff.Mode.CLEAR)
             strokeWidth = thickness
             strokeJoin = Paint.Join.ROUND
             strokeCap = Paint.Cap.ROUND
