+++ app/src/main/java/woowacourse/paint/customview/paint/tool/PaintTool.kt
@@ -1,4 +1,6 @@
-package woowacourse.paint.customview
+package woowacourse.paint.customview.paint.tool
+
+import woowacourse.paint.customview.paint.Painting
 
 sealed interface PaintTool {
     val painting: Painting
