+++ app/src/main/java/woowacourse/paint/main/ColorClickListener.kt
@@ -1,6 +1,6 @@
 package woowacourse.paint.main
 
-import woowacourse.paint.customview.BrushColor
+import woowacourse.paint.customview.paint.BrushColor
 
 interface ColorClickListener {
     fun onColorClick(brushColor: BrushColor)
