+++ app/src/main/java/woowacourse/paint/main/model/BrushColorBox.kt
@@ -1,6 +1,6 @@
 package woowacourse.paint.main.model
 
-import woowacourse.paint.customview.BrushColor
+import woowacourse.paint.customview.paint.BrushColor
 
 data class BrushColorBox(
     val brushColor: BrushColor,
