+++ app/src/main/res/layout/activity_main.xml
@@ -24,7 +24,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:titleTextColor="@color/white" />
 
-        <woowacourse.paint.customview.PaintingView
+        <woowacourse.paint.customview.paint.PaintingView
             android:id="@+id/pv_main"
             android:layout_width="match_parent"
             android:layout_height="0dp"
